pyproject.toml
[project]
name = "pizza-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.20.0.post0",
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
]
